% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{all_subclass_methods}
\alias{all_subclass_methods}
\title{All sub-class methods}
\usage{
all_subclass_methods(x)
}
\arguments{
\item{x}{Object with class}
}
\value{
Methods by sub-class, list of character vectors
}
\description{
All sub-class methods
}
\examples{
all_subclass_methods(0L)
}
