% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canny.tudor-package.R
\docType{package}
\name{canny.tudor-package}
\alias{canny.tudor}
\alias{canny.tudor-package}
\title{canny.tudor: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Roy Ratcliffe \email{royratcliffe@me.com} (\href{https://orcid.org/0000-0001-6594-1280}{ORCID})

}
\keyword{internal}
