% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cement.R
\name{cement_as_string}
\alias{cement_as_string}
\title{Cement argument names into list of strings}
\usage{
cement_as_string(...)
}
\arguments{
\item{...}{Arbitrary symbols}
}
\value{
List of strings
}
\description{
The arguments must not contain assignments, only symbols.
}
