% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{collect_column_classes_by_table}
\alias{collect_column_classes_by_table}
\title{Collect All Column Name and Class by Table}
\usage{
collect_column_classes_by_table(x, ...)
}
\arguments{
\item{x}{Data model, \code{dm} instance.}

\item{...}{
  Arguments passed on to \code{\link[=collect_name_and_class]{collect_name_and_class}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Collect All Column Name and Class by Table
}
