% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{collect_name_and_class}
\alias{collect_name_and_class}
\title{Collect Name and Class}
\usage{
collect_name_and_class(x, ...)
}
\arguments{
\item{x}{Data frame to collect.}

\item{...}{
  Arguments passed on to \code{\link[tibble:tibble]{tibble::tibble}}
  \describe{
    \item{\code{.rows}}{The number of rows, useful to create a 0-column tibble or
just as an additional check.}
    \item{\code{.name_repair}}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}
  }}
}
\value{
Data frame with \code{name} and \code{class} columns, each column of
character type.
}
\description{
Performs a zero-row collection on the incoming frame in order to obtain an
empty \code{\link[tibble]{tibble}}.
}
