% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{copy.env}
\alias{copy.env}
\title{Copy from one environment to another}
\usage{
copy.env(from.envir, to.pos = -1L, to.envir = as.environment(to.pos))
}
\arguments{
\item{from.envir}{Environment to copy from}

\item{to.pos}{\code{-1L} for the calling environment}

\item{to.envir}{Environment to copy to}
}
\description{
Copy from one environment to another
}
