% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{create_if_not_exists}
\alias{create_if_not_exists}
\title{Create scheme unless exists}
\usage{
create_if_not_exists(x, ...)
}
\arguments{
\item{x}{Schema to create.}

\item{...}{Extra arguments passed to method implementation.}
}
\description{
Create scheme unless exists
}
