% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{create_if_not_exists.schema}
\alias{create_if_not_exists.schema}
\title{Create scheme if it does not yet exist}
\usage{
\method{create_if_not_exists}{schema}(x, .con = DBI::ANSI(), ...)
}
\arguments{
\item{x}{Schema to create, if not existing}

\item{.con}{Database connect to use for literal quoting and for executing the
schema-glued statement}

\item{...}{
  Arguments passed on to \code{\link[DBI:dbExecute]{DBI::dbExecute}}
  \describe{
    \item{\code{conn}}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}
    \item{\code{statement}}{a character string containing SQL.}
  }}
}
\description{
Create scheme if it does not yet exist
}
