% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{db_drop}
\alias{db_drop}
\title{Drop database entity}
\usage{
db_drop(x, ...)
}
\arguments{
\item{x}{Entity to drop.}

\item{...}{Extra arguments passed to method implementation.}
}
\description{
Drop database entity
}
