% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{db_drop.schema}
\alias{db_drop.schema}
\title{Drop schema}
\usage{
\method{db_drop}{schema}(x, if_exists = FALSE, cascade = FALSE, .con = NULL, ...)
}
\arguments{
\item{x}{Schema to drop.}
}
\description{
Drop schema
}
