% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependentPackages}
\alias{dependentPackages}
\title{Dependent packages}
\usage{
dependentPackages(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[gtools:getDependencies]{gtools::getDependencies}}
  \describe{
    \item{\code{pkgs}}{character vector of package names}
    \item{\code{dependencies}}{character vector of dependency types to include.
Choices are "Depends", "Imports", "LinkingTo", "Suggests", and "Enhances".
Defaults to \code{c("Depends", "Imports", "LinkingTo")}.}
    \item{\code{installed}}{Logical indicating whether to pull dependency information
from installed packages. Defaults to TRUE.}
    \item{\code{available}}{Logical indicating whether to pull dependency information
from available packages.  Defaults to TRUE.}
    \item{\code{base}}{Logical indicating whether to include dependencies on base
packages that are included in the R installation. Defaults to FALSE.}
    \item{\code{recommended}}{Logical indicating whether to include dependencies on
recommended packages that are included in the R installation. Defaults to
FALSE.}
  }}
}
\value{
List of \code{\link[R.oo]{Package}} dependencies.
}
\description{
Finds \code{\link[R.oo]{Package}} objects for all dependencies of \code{pkgs}.
}
\examples{
\dontrun{
library(R.oo)
methods(class = "Package")
#  [1] as.character        getAuthor           getBundle           getBundlePackages
#  [5] getChangeLog        getClasses          getContents         getContribUrl
#  [9] getDataPath         getDate             getDescription      getDescriptionFile
# [13] getDevelUrl         getDocPath          getEnvironment      getExamplePath
# [17] getHistory          getHowToCite        getLicense          getMaintainer
# [21] getName             getNews             getPath             getPosition
# [25] getTitle            getUrl              getVersion          isLoaded
# [29] isOlderThan         ll                  load                showChangeLog
# [33] showContents        showDescriptionFile showHistory         showHowToCite
# [37] showNews            startupMessage      unload

dependentPackages("dplyr") |> tibble::enframe(value = "package")
}
}
