% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_get_all_pk_cols}
\alias{dm_get_all_pk_cols}
\title{Get All Primary Key Columns}
\usage{
dm_get_all_pk_cols(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=dm_get_all_pks]{dm_get_all_pks}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Data frame of \code{table} and \code{pk_cols}; the latter collecting
a list-wrapped vector of sorted primary-key columns.
}
\description{
Get All Primary Key Columns
}
