% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_get_all_pks}
\alias{dm_get_all_pks}
\title{Get All Primary Keys from Data Model}
\usage{
dm_get_all_pks(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[dm:dm_get_all_pks]{dm::dm_get_all_pks}}
  \describe{
    \item{\code{dm}}{A \code{dm} object.}
    \item{\code{table}}{One or more table names, as character vector,
to return primary key information for.
The default \code{NULL} returns information for all tables.}
  }}
}
\value{
Data frame of two variables: \code{table} of character and
\code{pk_col} for the primary-key column, also of character.
}
\description{
Gets all keys but also unlists the \code{pk_col} column.
}
