% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_system_info}
\alias{docker_system_info}
\title{System-wide Docker information}
\usage{
docker_system_info()
}
\value{
List of system-wide information pairs. There are some 60 pairs,
typically.
}
\description{
System-wide Docker information
}
