% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_config}
\alias{.docker_config}
\title{Home directory's docker configuration}
\usage{
.docker_config()
}
\description{
Begins with full-stop (dot) in order not to overlap "docker config" as a
command. The latter manages Docker configurations, while this function reads
only the calling user's configuration file.
}
