% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named-list.R
\name{flatten_named.list}
\alias{flatten_named.list}
\title{Flatten a named list}
\usage{
flatten_named.list(x)
}
\arguments{
\item{x}{Named list.}
}
\value{
Unnamed list of names and values, interlaced.
}
\description{
Named list to interlaced names and values, each name followed
immediately by its associated value.
}
