% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{get.gitlab.blobs}
\alias{get.gitlab.blobs}
\title{Gets project blobs}
\usage{
get.gitlab.blobs(x, project = character(1L), ...)
}
\arguments{
\item{x}{Project files with path column.}

\item{project}{Identifies which project to apply file paths to.}

\item{...}{
  Arguments passed on to \code{\link[gitlabr:gl_repository]{gitlabr::gl_get_file}}
  \describe{
    \item{\code{ref}}{name of ref (commit branch or tag)}
    \item{\code{file_path}}{path to file}
    \item{\code{to_char}}{flag if output should be converted to char; otherwise it is of class raw}
    \item{\code{api_version}}{a switch to force deprecated GitLab API v3 behavior. See details section "API version" of \code{\link[gitlabr:gl_connection]{gl_connection()}}}
  }}
}
\description{
Encodes the file paths using \code{\link[utils]{URLencode}}.
}
