% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{get.gitlab.project.id}
\alias{get.gitlab.project.id}
\title{Gets project identifier as integer}
\usage{
get.gitlab.project.id(...)
}
\description{
You can use the extra arguments to filter the incoming project tibble by path or other criteria.
}
\examples{
\dontrun{
list.gitlab.maintainer.projects() \%>\%
  get.gitlab.project.id(path == "path-of-gitlab-project")
}
}
