% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{glue_sql}
\alias{glue_sql}
\title{Glue SQL schema}
\usage{
glue_sql(x, ...)
}
\arguments{
\item{x}{Schema to use for gluing.}

\item{...}{Extra arguments passed to method implementation.}
}
\description{
Glue SQL schema
}
