% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{glue_sql.schema}
\alias{glue_sql.schema}
\title{Glue schema to SQL statement}
\usage{
\method{glue_sql}{schema}(x, ...)
}
\arguments{
\item{x}{Schema to use for gluing.}

\item{...}{
  Arguments passed on to \code{\link[glue:glue_sql]{glue::glue_sql}}
  \describe{
    \item{\code{.con}}{[\code{DBIConnection}]:A DBI connection object obtained from \code{DBI::dbConnect()}.}
    \item{\code{.envir}}{[\code{environment}: \code{parent.frame()}]\cr Environment to evaluate each expression in. Expressions are
evaluated from left to right. If \code{.x} is an environment, the expressions are
evaluated in that environment and \code{.envir} is ignored. If \code{NULL} is passed it is equivalent to \code{\link[=emptyenv]{emptyenv()}}.}
    \item{\code{.na}}{[\code{character(1)}: \sQuote{NA}]\cr Value to replace NA values
with. If \code{NULL} missing values are propagated, that is an \code{NA} result will
cause \code{NA} output. Otherwise the value is replaced by the value of \code{.na}.}
  }}
}
\description{
Glue schema to SQL statement
}
