% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx.R
\name{hdx}
\alias{hdx}
\title{Half-duplex communication}
\usage{
hdx(x, ...)
}
\arguments{
\item{x}{Socket connection.}

\item{...}{
  Arguments passed on to \code{\link[=hdx.sockconn]{hdx.sockconn}}
  \describe{
    \item{\code{text}}{Lines of text to write to the socket connection, a character
vector. Writes all the lines before reading.}
    \item{\code{write.sep}}{Line separator written after every line of \code{text}, a
character string defaulting to a single new-line character \code{"\n"}.}
    \item{\code{write.useBytes}}{Suppresses re-encoding of marked strings.}
    \item{\code{timeout}}{Seconds to wait for ready socket, a numeric; or wait
indefinitely, \code{NULL}.}
  }}
}
\description{
Half-duplex communication
}
