% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{list.gitlab.maintainer.projects}
\alias{list.gitlab.maintainer.projects}
\title{Lists maintainer-level projects using default GitLab connection}
\usage{
list.gitlab.maintainer.projects(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[gitlabr:gl_list_projects]{gitlabr::gl_list_projects}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Lists maintainer-level projects using default GitLab connection
}
