% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{list.gitlab.project.files}
\alias{list.gitlab.project.files}
\title{Lists tree recursively}
\usage{
list.gitlab.project.files(...)
}
\arguments{
\item{...}{Arguments for gl_list_files}
}
\description{
Runs iteratively for all tree-type files. The tree path becomes a new path for a sub-listing.
}
