% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{p.q.frame}
\alias{p.q.frame}
\title{Permutations of p-q pairs.}
\usage{
p.q.frame(v = vector())
}
\arguments{
\item{v}{Source vector.}
}
\value{
Permutation frame.
}
\description{
Permutes a pair of variables, p and q, by a data frame such that each frame
row contains a unique p-q permutation.
}
\examples{
p.q.frame(c("a", "b"))
#>   p q
#> 1 a b
#> 2 b a

p.q.frame(1:3)
#>   p q
#> 1 1 2
#> 2 1 3
#> 3 2 1
#> 4 2 3
#> 5 3 1
#> 6 3 2
}
