% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{reparse_url}
\alias{reparse_url}
\title{Parse and re-parse while only scheme and path}
\usage{
reparse_url(x)
}
\arguments{
\item{x}{URL to parse for components.}
}
\value{
URL components with vectorised nested scheme.
}
\description{
Parses the given URL and continuously parses the resulting path so long as
the result is only a scheme and path. This occurs for nested schemes. The
final URL accumulates the scheme as a character vector with super- to
sub-scheme ordering.
}
\details{
The implementation repeatedly parses, that is re-parses, while non-null URL
components result in only a scheme-path pair. Note that URLs carry NULL
values for non-existent URL components.
}
\examples{
url <- reparse_url("superscheme:subscheme://host/path")
url$scheme
# [1] "superscheme" "subscheme"
}
