% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{set.gitlab.connection}
\alias{set.gitlab.connection}
\title{Sets up a connection to gitlab.com using GITLAB_PRIVATE_TOKEN}
\usage{
set.gitlab.connection(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[gitlabr:gl_connection]{gitlabr::gl_connection}}
  \describe{
    \item{\code{gitlab_url}}{URL to the GitLab instance (e.g. \verb{https://gitlab.myserver.com})}
    \item{\code{private_token}}{private_token with which to identify. You can generate one in the web interface under
\verb{GITLABINSTANCEURL/-/profile/personal_access_tokens.html} when logged on.}
    \item{\code{api_version}}{Currently "4" for the latest GitLab API version. See Details section on API versions.}
    \item{\code{api_location}}{location of the GitLab API under the \code{gitlab_url}, usually and by default "/api/${api_version}/"}
  }}
}
\description{
Sets up a connection to gitlab.com using GITLAB_PRIVATE_TOKEN
}
