% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlformat.R
\name{sql.format}
\alias{sql.format}
\title{Formats SQL using the sqlformat.org API.}
\usage{
sql.format(sql, ...)
}
\arguments{
\item{sql}{Structured Query Language.}

\item{...}{Format parameters.}
}
\description{
See example usage below. Arguments
are named list elements corresponding to format parameters. Only \code{sql} is
required; the rest have sensible defaults.

See https://sqlformat.org/api/ for details. A usage limit exists: 500 requests
per hour based on the IP address.
}
\examples{
sql.format(sql="select * from foo", reindent=1, keyword_case="upper")
# -> [1] "SELECT *\nFROM foo"
}
