% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{strsplit.trimws}
\alias{strsplit.trimws}
\title{Split string and trim white space}
\usage{
strsplit.trimws(x, split, ...)
}
\arguments{
\item{x}{What to split, a character vector.}

\item{split}{What to use for splitting. Either a character vector or an
object that can coerce to such.}

\item{...}{
  Arguments passed on to \code{\link[base:strsplit]{base::strsplit}}
  \describe{
    \item{\code{fixed}}{
    logical.  If \code{TRUE} match \code{split} exactly, otherwise
    use regular expressions.  Has priority over \code{perl}.
  }
    \item{\code{perl}}{logical.  Should Perl-compatible regexps be used?}
    \item{\code{useBytes}}{logical.  If \code{TRUE} the matching is done
    byte-by-byte rather than character-by-character, and inputs with
    marked encodings are not converted.  This is forced (with a warning)
    if any input is found which is marked as \code{"bytes"}
    (see \code{\link[base]{Encoding}}).}
  }}
}
\value{
Split strings, a character vector.
}
\description{
Trims both leading and trailing white space.
}
