% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Sys.java_home}
\alias{Sys.java_home}
\title{Java home based on Java binary}
\usage{
Sys.java_home(bin_java = Sys.which("java"))
}
\arguments{
\item{bin_java}{Path to Java binary.}
}
\description{
Compute \code{JAVA_HOME} based on the current Java path. Look for Java in the
search path. Strip away its super-folders. Assume and assert that the first
super-folder matches \code{bin}. Fail if it does \emph{not} match.
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{bin.java <- Sys.which("java")
home.bin <- dirname(bin.java)
stopifnot(basename(home.bin) == "bin")
java.home <- dirname(home.bin)
}\if{html}{\out{</div>}}

The \code{bin} super-folder's parent is Java \emph{home}.
}
