% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Sys.setenv_java_home}
\alias{Sys.setenv_java_home}
\title{Set Java home based on Java binary}
\usage{
Sys.setenv_java_home(java_home = Sys.java_home())
}
\arguments{
\item{java_home}{New path for \code{JAVA_HOME}.}
}
\description{
Stops if \code{java_home} is not a currently existing directory.
}
