% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{types.of}
\alias{types.of}
\title{Types of tibble or data frame}
\usage{
types.of(x)
}
\arguments{
\item{x}{Tibble or data frame}
}
\value{
Named types, vector of character
}
\description{
Types of tibble or data frame
}
