% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.R
\name{upload_feather}
\alias{upload_feather}
\title{Upload data frame as feather}
\usage{
upload_feather(x, ...)
}
\arguments{
\item{x}{Data frame to upload}

\item{...}{Extra arguments for \code{httr::\link[httr]{POST}}}
}
\description{
Utilises \link{httr}, \link{feather} and \link{withr} packages.
}
