% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdx.R
\name{vhdx.sockconn}
\alias{vhdx.sockconn}
\title{Vectorised half-duplex socket communication}
\usage{
\method{vhdx}{sockconn}(x, text, write.sep = "\\n", write.useBytes = FALSE, timeout = NULL, ...)
}
\description{
Vectorised half-duplex socket communication
}
