% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vunzip_bin}
\alias{vunzip_bin}
\title{Vectorised unzip to raw vectors}
\usage{
vunzip_bin(x, ...)
}
\arguments{
\item{x}{Vector of zip files either by name or by raw vector.}

\item{...}{
  Arguments passed on to \code{\link[=unzip_bin]{unzip_bin}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Zipper list of member lists of raw vector. Does \emph{not} attempt to
simplify the resulting list of lists of raw vectors. Consistency makes this
useful since sub-lists can easily have different file names.
}
\description{
Vectorised unzip to raw vectors
}
\examples{
# Note the trailing dollar in the pattern below. Only matches path names
# ending with `zip` extension.
zippers <- canny.tudor::vunzip_bin(list.files(pattern = "*.zip$"))

# This is how you might iterate the resulting list of lists, of zippers of
# members, using nested `for` loops.
for (i in seq_along(zippers)) {
  members <- zippers[[i]]
  for (j in seq_along(members)) {
    cat(paste(names(zippers)[[i]], names(members)[[j]], sep = ":"),
        length(members[[j]]), "\n")
  }
}
}
