% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{Windows.registry_path}
\alias{Windows.registry_path}
\title{Windows Registry Path}
\usage{
Windows.registry_path(...)
}
\arguments{
\item{...}{Registry path components, one or more.}
}
\value{
Back-slash delimited registry path.
}
\description{
Constructs a Windows registry path using back-slash delimiters. The default
file path separator is \emph{not} back-slash on Windows, but forward-slash
instead.
}
