% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{collect_table_column_classes}
\alias{collect_table_column_classes}
\title{Collects Classes of Table Columns}
\usage{
collect_table_column_classes(x, ...)
}
\arguments{
\item{x}{\code{dm} class object, a data model.}

\item{...}{Ignored parameters.}
}
\description{
Collects Classes of Table Columns
}
