% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{copy.env}
\alias{copy.env}
\title{Copy from one environment to another}
\usage{
copy.env(from.env, to.env)
}
\arguments{
\item{from.env}{Environment to copy from}

\item{to.env}{Environment to copy to}
}
\description{
Copy from one environment to another
}
