% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{get.gitlab.blobs}
\alias{get.gitlab.blobs}
\title{Gets project blobs}
\usage{
get.gitlab.blobs(x, project = character(1L), ...)
}
\arguments{
\item{x}{Project files with path column.}

\item{project}{Identifies which project to apply file paths to.}
}
\description{
Encodes the file paths using \code{\link[utils]{URLencode}}.
}
