% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{list.gitlab.owner.projects}
\alias{list.gitlab.owner.projects}
\title{Lists owner-level projects using default GitLab connection}
\usage{
list.gitlab.owner.projects(...)
}
\description{
Lists owner-level projects using default GitLab connection
}
