% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{list.gitlab.project.files}
\alias{list.gitlab.project.files}
\title{Lists tree recursively}
\usage{
list.gitlab.project.files(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[gitlabr:gl_repository]{gitlabr::gl_list_files}}
  \describe{
    \item{\code{project}}{name or id of project (not repository!)}
    \item{\code{ref}}{name of ref (commit branch or tag)}
  }}
}
\value{
Tibble of project files comprising tree and blob types.
}
\description{
Runs iteratively for all tree-type files. The tree path becomes a new path
for a sub-listing.
}
