% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{set.gitlab.connection}
\alias{set.gitlab.connection}
\title{Sets up a connection to gitlab.com using GITLAB_PRIVATE_TOKEN}
\usage{
set.gitlab.connection(...)
}
\description{
Sets up a connection to gitlab.com using GITLAB_PRIVATE_TOKEN
}
